AD: added REST API functionality to PersonRegister page

+++ src/components/api/person/model/PersonCard.js
@@ -1,4 +1,5 @@
 import { useState } from "react";
+import OverlayMenu from "../../Overlay";
 import PersonForm from "./ui/PersonForm";
 
 const PersonCard = ({personObj, onEdit, onDelete}) => {
@@ -11,16 +12,19 @@ const PersonCard = ({personObj, onEdit, onDelete}) => {
     }
 
     const deleteEntity = () => {
-        onDelete(personObj.id);
+        if(window.confirm("Are you sure you wish to delete: " + personObj.forename + " " + personObj.surname+"?")){
+            onDelete(personObj.id);
+        }
     }
 
     return  editMode ? 
-            ( <div key={personObj.id} className="personCard"> <h3>{personObj.id}</h3> <PersonForm action="Edit" onSubmit={editEntity} /> </div>)
+            ( <div key={personObj.id} className="personCard"> <h3>{personObj.id}</h3> 
+            <PersonForm action="Edit" onSubmit={editEntity} onCancel={() => setEdit(false)} forename={personObj.forename} surname={personObj.surname} age={personObj.age} /> </div>)
             :
             (
             <div key={personObj.id} className="personCard">
                 <div className="personOptions">
-                <input className="btn" type="button" value="Edit" onClick={editEntity}/>
+                <input className="btn" type="button" value="Edit" onClick={()=>setEdit(true)} />
                 <input className="btn"  type="button" value="Delete" onClick={deleteEntity}/>
 
                 </div>
